requirements.txt
requests>=2.30.0
selenium>=4.37.0
python-dotenv>=1.0.1
pandas>=2.2.3
undetected-chromedriver>=3.5.5